pyproject.toml
[project]
name = "financier-bot"
version = "0.1.0"
description = "Telegram expense bot with FastAPI backend and Google Sheets integration"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Artur", email = "romanchenkoas@yandex.ru" }
]
license = { text = "MIT" }

dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "httpx>=0.27.0",
  "loguru>=0.7.2",
  "aiogram>=3.4.1",
  "gspread>=6.0.0",
  "google-auth>=2.30.0",
  "tenacity>=8.3.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "python-dotenv>=1.0.0",
  "black>=25.9.0",
  "isort>=6.0.1",
]

[project.scripts]
financier-bot = "src.telegram.bot:run"

[tool.uv]
# uv uses standard PEP 621 metadata; no special config required
package = true

[tool.black]
line-length = 100
